requirements.txt
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.3
openai==1.9.0
pandas==2.2.0
python-dotenv==1.0.0
scipy==1.12.0
seaborn==0.13.1
tiktoken==0.5.1
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2